requirements.txt
streamlit
pydub
yt-dlp
openai-whisper
transformers
torch
